requirements.txt
fastapi
uvicorn[standard]
motor
dnspython
pydantic[email]
python-dotenv
bcrypt==3.2.2
python-multipart
python-jose[cryptography]
passlib[bcrypt]
apscheduler
